+++ changelog.md
@@ -1,4 +1,5 @@
 HEAD
+- Allows changing the listen backlog queue length.
 - Fix handler allocation crash with multithreaded io_service.
 - Split tcp init into pre and post init.
 - Adds URI method to extract query string from URI. Thank you Banaan for code.